Mmsapt: rewrite TSA_tree_iterator::getPid() to return a phrase hash
Since we are going to use multiple indices, the current definition of phrase ID (sid<<32) + (off<<16) + (p+1) is not going to work anymore, because matches of the same phrase may occur in any domain.

Therefore, switch phrase ID to something that is consistent across domain corpora: the hash of the phrase word ID sequence.